requirements.txt
-i https://pypi.org/simple
-e .
-r test_requirements.txt
agate-dbf==0.2.1
agate-excel==0.2.3
agate-sql==0.5.4
agate==1.6.1
attrs==19.3.0
babel==2.7.0
bidict==0.18.3
boto3==1.10.28
botocore==1.13.28
cchardet==2.1.5
certifi==2019.11.28
chardet==3.0.4
click-default-group==1.2.2
click==7.0
csvkit==1.0.4
datapackage==1.10.0
dateparser==0.7.2
dbfread==2.0.7
decorator==4.4.1
docutils==0.15.2
et-xmlfile==1.0.1
goodtables==2.4.1
idna==2.8
ijson==2.5.1
importlib-metadata==0.23 ; python_version < '3.8'
isodate==0.6.0
jdcal==1.4.1
jinja2==2.10.3
jmespath==0.9.4
jsonlines==1.2.0
jsonpointer==2.0
jsonschema==3.2.0
leather==0.3.3
linear-tsv==1.1.0
markupsafe==1.1.1
more-itertools==7.2.0
openpyxl==3.0.2
pandas==0.25.3
parsedatetime==2.5
pyrsistent==0.15.6
python-dateutil==2.8.0 ; python_version >= '2.7'
python-magic==0.4.15
python-slugify==4.0.0
pytimeparse==1.1.8
pytz==2019.3
regex==2019.11.1
requests==2.22.0
rfc3986==1.3.2
s3transfer==0.2.1
simpleeval==0.9.8
six==1.13.0
sqlalchemy==1.3.11
statistics==1.0.3.5
tableschema==1.11.0
tabulator==1.30.0
text-unidecode==1.3
tzlocal==2.0.0
unicodecsv==0.14.1
urllib3==1.25.7 ; python_version >= '3.4'
validators==0.14.0
xlrd==1.2.0
xlsx2csv==0.7.6
zipp==0.6.0
